Refactor snaps-execution-environments build to worker threads
Refactored snaps-execution-environments build to use Worker threads to enable multi-threaded bundle building. This change will speed up the build process.
